+++ rust/operator-binary/src/controller/validate.rs
@@ -153,6 +153,7 @@ pub fn validate(
         uid,
         cluster.spec.nodes.role_config.clone(),
         role_group_configs,
+        cluster.spec.cluster_config.tls.clone(),
         cluster.spec.cluster_config.keystore.clone(),
     ))
 }
@@ -197,6 +198,7 @@ fn validate_role_group_config(
         listener_class: merged_role_group.config.config.listener_class,
         logging,
         node_roles: merged_role_group.config.config.node_roles,
+        requested_secret_lifetime: merged_role_group.config.config.requested_secret_lifetime,
         resources: merged_role_group.config.config.resources,
         termination_grace_period_seconds,
     };
@@ -295,7 +297,8 @@ mod tests {
             role_utils::{GenericProductSpecificCommonConfig, RoleGroupConfig},
             types::{
                 kubernetes::{
-                    ConfigMapName, ListenerClassName, NamespaceName, SecretKey, SecretName,
+                    ConfigMapName, ListenerClassName, NamespaceName, SecretClassName, SecretKey,
+                    SecretName,
                 },
                 operator::{
                     ClusterName, ControllerName, OperatorName, ProductName, ProductVersion,
@@ -425,6 +428,8 @@ mod tests {
                                 ]
                                 .into()
                             ),
+                            requested_secret_lifetime: Duration::from_str("1d")
+                                .expect("should be a valid duration"),
                             resources: Resources {
                                 memory: MemoryLimits {
                                     limit: Some(Quantity("2Gi".to_owned())),
@@ -510,6 +515,10 @@ mod tests {
                     }
                 )]
                 .into(),
+                v1alpha1::OpenSearchTls {
+                    server_secret_class: Some(SecretClassName::from_str_unsafe("tls")),
+                    internal_secret_class: SecretClassName::from_str_unsafe("tls")
+                },
                 vec![v1alpha1::OpenSearchKeystore {
                     key: OpenSearchKeystoreKey::from_str_unsafe("Keystore1"),
                     secret_key_ref: v1alpha1::SecretKeyRef {
@@ -694,6 +703,7 @@ mod tests {
                 image: serde_json::from_str(r#"{"productVersion": "3.1.0"}"#)
                     .expect("should be a valid ProductImage structure"),
                 cluster_config: v1alpha1::OpenSearchClusterConfig {
+                    tls: v1alpha1::OpenSearchTls::default(),
                     keystore: vec![v1alpha1::OpenSearchKeystore {
                         key: OpenSearchKeystoreKey::from_str_unsafe("Keystore1"),
                         secret_key_ref: v1alpha1::SecretKeyRef {